feat:  SelectedMemo 를 싱글톤으로 변경
